requirements.txt
docopt
feedparser
pyyaml
